LUA implementation relocations.
