+++ src/celengine/command.cpp
@@ -12,9 +12,9 @@
 #include "asterism.h"
 #include "execution.h"
 #include "glcontext.h"
-#include <celestia/celestiacore.h>
+#include <celestia/CelestiaCoreApplication.h>
 #include <celestia/imagecapture.h>
-#include <celestia/celx_internal.h>
+#include <celestia/celx_lua.h>
 #include <celutil/util.h>
 #include <iostream>
 #include "eigenport.h"
